+++ sample/publish.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+eval $(opam env)
+gh-pages-amend $(odig cache path)/html doc